Interlink Admin Office (#1015)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Makes a proper office for Centcom Administration on the Interlink.
Includes a small bedroom for ERP as well.
Essentially, I've been messing around with my amazing admin powers a lot
and I figured: "Hey. Why do we still do stuff from the old Centcom area
when we bluespace-spawn? The Interlink was mean to make the connection
between Centcom and the Station closer ever since [other server] decided
'nah fuck the old centcom building'. So here we are, a cute little
office." Thus, the cute little Administrative Office was born. Yes, it's
called the Administrative Office as a joke since admins will use it.


![dreamseeker_2024-01-15_17-51-18](https://github.com/Bubberstation/Bubberstation/assets/5521322/c52e43b1-84e1-4ef1-b272-7021ff13db6f)


## Why It's Good For The Game

Good for the game? It's an admeme map edit! Now the old CentCom map is
even more obsolete!

## Proof Of Testing

See included image.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Centcom Administration has moved to the Interlink
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->