Nerfs most modsuits by giving them all armour boosters (#910)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

All security-grade modsuits (security, safeguard, magnate, blueshield,
ERT, etc) have been outfitted with armour booster modules, built in and
non-removable.

"But wait, Yakumo, that's a huge buff, isn't it?"

haha no

In """"compensation"""" for this built-in module, I have halved the
armour values on all applicable modsuits, and moved said values to the
armour booster modules. Every modsuit has its own specific booster with
the moved armour values built in.

Also almost all of them don't get the speed boost, because it's just a
poor copy of the syndicate one. I made an exception for the HoS MOD
however because the secoff modsuits were recently buffed to have almost
the same armour (we love power creep)


![image](https://github.com/Bubberstation/Bubberstation/assets/10399117/92718467-562f-4ffb-afae-85d01d5d5761)
 
Contains an early pull of
https://github.com/tgstation/tgstation/pull/80680

## Why It's Good For The Game

EVA combat is incredibly goofy right now. Armour on MODsuits was
originally stripped out severely because having armour in already awful
space combat was a little bit shit, and modsuits aren't supposed to be a
be-all-end-all of protective drip. Except the Syndicate who got shafted
by having dogshit armour in space, because most of their armour is in
the armour booster.

Now everyone is on the same playing field again. Want to have actual
armour in a MODsuit? You're gonna have to sacrifice your EVA capability
to do so. No more free rides, and the syndicate is a respective threat
in comparison in the same EVA environment again.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Security, HoS, Captain, Blueshield, and Centcom roles have had
armour booster modules built into their modsuits. Most do not reduce the
slowdown of your suit (except the HoS one)
balance: Armour values in said modsuits have been halved across the
board, with the lost values moved to the armour boosters.
balance: Effectively, armour in EVA combat is severely nerfed for
everyone except the syndicate who already had armour boosters.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>
Co-authored-by: The Sharkening <95130227+StrangeWeirdKitten@users.noreply.github.com>