[TESTMERGE] Bloodsuckers, version 3 (#863)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds Fulpstations bloodsuckers into the game, this is based on
https://github.com/Bubberstation/Bubberstation/pull/386 which also has
good info on the abilities that it can do, although fulp also has their
own wiki page on it, https://wiki.fulp.gg/en/Bloodsucker.
This was built upon from the last PR and with the major known issues
fixed.
Balance changes:

- Bloodsucker snow need 0.4 of your max blood level for each level up
-  Bloodsuckers can no longer stockpile more than two level ups.
- Draining mindless humans gives the same draining inefficiency as
draining monkeys does, which gets twice as bad if you are above level 4.
- If you are at level 4, you inherit the snobbish blood drinking quality
just like venture always have, with the exception of the nosferatu. This
means you'll refuse to drink bad blood, to encourage conflict if you
desire even more power.
- Frenzy no longer kills you as quickly, tested it and it was incredibly
fast.
- Night timer for vampires is now 20 minutes.
- All vampire type species and bloodsuckers get hurt more by silver

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
New antags add variety and fun. Probably
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Add bloodsuckers, a new reason to fear maintenance and the dark.
Don't go out there alone, or you might wake up pale and fainting.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: destrucktoid <39823128+destrucktoid@users.noreply.github.com>
Co-authored-by: Ldip999 <60290575+Ldip999@users.noreply.github.com>
Co-authored-by: StrangeWeirdKitten <95130227+StrangeWeirdKitten@users.noreply.github.com>