REDSEC IS OVER... (#1047)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
https://www.youtube.com/watch?v=6kguaGI7aZg

Removes the automapped redsec vendor

## Why It's Good For The Game

#986 has eclipsed the need for a second, identical (except sprite)
vendor in the security office.

:cl:
del: After the introduction of the Multisec Vendor, stations are no
longer shipped with an additional Redsec vendor.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>