[Continues #879] Adds corrupt hounds as miner borgs (#1029)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
This adds new wideborgs to the game: The Corrupt borg and Corrupt alt
borgs, as Miner sprites. Big and scary, these guys look like they bite.
<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
More variety is always good for the game! Now miner borgs as part of
sprite alignment.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing

![image](https://github.com/Bubberstation/Bubberstation/assets/110273561/34009d3f-73d4-49a6-87f5-cece71962f8c)

![image](https://github.com/Bubberstation/Bubberstation/assets/110273561/879f5de9-dd50-4c72-a7d7-75fd78bf8118)

![image](https://github.com/Bubberstation/Bubberstation/assets/110273561/075b5893-79c5-47fb-8411-b7cb33811728)

<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

🆑 ReturnToZender
add: Corrupt and Corrupt Alt borg sprites to the peacekeeper tab!
fix: code cleanup in the borg modular file, fixed a typo, added a
missing mapload
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Esixdotnet <156951974+Esixdotnet@users.noreply.github.com>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>
Co-authored-by: The Sharkening <95130227+StrangeWeirdKitten@users.noreply.github.com>
Co-authored-by: Bubberbot <151680451+Bubberbot@users.noreply.github.com>
Co-authored-by: thegrb93 <grbrown93@sbcglobal.net>
Co-authored-by: BurgerLUA <8602857+BurgerLUA@users.noreply.github.com>