Replaces the Oh Hi Daniel shuttle with something else (#1071)
## About The Pull Request


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/c58bc964-b052-4947-9ef1-b5c52b025be9)

Why yes, I watched several of the same clip of the Room several times to
get the layout just right.

## Why It's Good For The Game

The Oh Hi Daniel shuttle is much more chaotic on here than it is on /tg/
due to the fact that anyone can spam alcohol at the vendor and then
flood the entire shuttle with very flammable liquid. Causes a lot of
headaches for admins, so I decided to remove that aspect for it and then
pretty much change it completely because I'm legitimately wild.


## Proof Of Testing

Draft because untested.

## Changelog

:cl: BurgerBB
add: Replaces the Oh Hi Daniel shuttle with a similar shuttle.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>