[no longer a draft] Digi cyberlegs (#973)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

Digitigrade cybernetic legs. Advanced/surplus are not complete
sprite-wise so use standard icons as placeholder. This is part 1 in an
effort to fix [Issue
#705](https://github.com/Bubberstation/Bubberstation/issues/705)

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

MORE cybernetic limb options. these never hurt anyone. except the
clutter in the augmentations+ leg section.

## Proof Of Testing

<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->


![Screenshot_64](https://github.com/Bubberstation/Bubberstation/assets/94389683/526566ff-1bdc-4dfe-849d-ecd4295707bf)
![Screenshot_63](https://github.com/Bubberstation/Bubberstation/assets/94389683/0083c0d4-5994-490f-94fd-537b77346db8)
![Screenshot_62](https://github.com/Bubberstation/Bubberstation/assets/94389683/440e845a-5e2f-4523-92b6-94a8f507c34e)
![Screenshot_61](https://github.com/Bubberstation/Bubberstation/assets/94389683/bbc1a30f-3736-42bc-a314-b74b960abd3a)


![Screenshot_73](https://github.com/Bubberstation/Bubberstation/assets/94389683/d7a38f76-3bd0-4925-a94e-fa979d562c25)![Screenshot_72](https://github.com/Bubberstation/Bubberstation/assets/94389683/f42c1c5e-cbb8-4a1b-9621-910ebf85e08a)
![Screenshot_71](https://github.com/Bubberstation/Bubberstation/assets/94389683/5295d69c-fd61-4304-9e13-47783b5926d1)
![Screenshot_70](https://github.com/Bubberstation/Bubberstation/assets/94389683/4d5593af-6dcd-487e-849f-6e9aee918440)

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: digitigrade cybernetic legs, and the ability to select them in
preferences
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>