Adds the xenomorph plushie (#1072)
## About The Pull Request

We have a lot of plushies here on Bubberstation 13. Many plushies.
Imagine my shock when I discover we do not have a xenomorph drone
plushie! Unforgivable, I say. This PR corrects it by making the little
beanomorph of a xeno plushie a thing. It can be selected in the loadout
menu, and makes the classic xenomorph death-hiss when squeaked.

EDIT: An inhand sprite has also been added, along with a maid variant,
also selectable in the loadout menu.


![plushxeno](https://github.com/Bubberstation/Bubberstation/assets/94123167/f0f4818e-25c3-42e9-834e-990e258f335a)

![plushxeno](https://github.com/Bubberstation/Bubberstation/assets/94123167/50eb7ee8-88bf-4d45-80e8-7dcf1df0e5b0)

## Why It's Good For The Game

The xenomorph is an iconic character in-game and out of game, and it
only feels right that it have a plushie too. Just look at its happy lil'
smile, it wants to be your friend!

## Proof Of Testing


![Untitled](https://github.com/Bubberstation/Bubberstation/assets/94123167/155746ba-224c-4934-b2e8-2151285d32fc)

## Changelog

:cl:

add: Added a xenomorph plushie.
add: Added a xenomorph maid plushie.

:cl:

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>