Adds Miasma Organ Rotting (#860)
## About The Pull Request

Organs that rot now create miasma. 

Note that miasma emissions will stop once the organ is fully decayed,
and the amount decayed per second is based on the damage dealt to
organs.

Roughly the amount of miasma created from one fully-decayed corpse will
be about 30 moles of miasma. For reference, a 1x1 room contains 104
moles of oxygen+nitrogen. During testing, I murdered and left Pun Pun to
rot for a bit in a 4x3 backroom. Miasma levels were about 2.2% until
they were about 70% rotted.

## Why It's Good For The Game

Miasma is hardly used as a gas in gameplay, and needs to honestly exist
in more places. Also, there needs to be a way to locate corpses easier
that are stuffed in lockers or in small rooms to prevent soft round
removal where a body is left in a place for ages and no one notices.

With the nerf to miasma in #649, the effects of miasma exposure
shouldn't be as bad as it was before.

## Changelog

:cl: BurgerBB
add: Organs that rot now create miasma. Note that miasma emissions will
stop once the organ is fully decayed.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>