Sets Evil Seedling TC cost to 2 (From 8), removes the ability for Seedlings to have multiple harvests. (#1043)
## About The Pull Request

Sets Evil Seedling TC cost to 2 (From 8), removes the ability for
Seedlings to have multiple harvests.

## Why It's Good For The Game

The last time someone made an army of seedlings, the station ended up
having 107% time dilation due to pathing/garbage collection runtimes.

The second last time someone made an army of seedlings, the person got
an antag permaban for wiping out the entirety of security and storming
command.

I think the maximum seedlings someone should have at any given time is
four seedlings, so this is why the TC cost was divided by 4. Of course,
traitors can buy more seedlings if they want to but this will put a dent
in their TC budget.

## Proof Of Testing

It came to me in a dream.

## Changelog

:cl: BurgerBB
del:  Removes the ability for Seedlings to have multiple harvests.
balance: Sets Evil Seedling TC cost to 2 (From 8),
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->