Nerfs Crab17 because I lost my life savings once (#865)
## About The Pull Request

The Crab17 uplink item cannot spawn in surplus crates, nor can it be
discounted.

## Why It's Good For The Game

The Crab17, while a fun addition, is kind of getting annoying fast
(nearly purchased every round to the point where people unironically
metagame it if they know who is traitor). To discourage overusage, it
cannot spawn in surplus bundles, nor can it be discounted.


## Changelog

:cl: BurgerBB
balance: The Crab17 uplink item cannot spawn in surplus crates, nor can
it be discounted.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>
Co-authored-by: The Sharkening <95130227+StrangeWeirdKitten@users.noreply.github.com>