Revert "Better airlock sounds QOL" (#1101)
Reverts Bubberstation/Bubberstation#1054

These are some of the most grating sounds I've ever heard added to the
game. If it was something we only heard a few times for round it'd be
fine, but it's EVERY AIRLOCK, one of the devices we interact with THE
MOST as players. Some sounds are extremely oud, some are quiet, some are
fine and some are grating like nails on a chalkboard. There's also a lot
of ones that are obscenely long for how simple and quick of an action
airlock opening and closing is.

It's an extremely inconsistent PR and should have been TM'd longer
before changing stuff players touch so often. None of the feedback from
the one TM was allowed to be acted on.


___________________________________________________________________________________________

:cl:
del: Door sounds from Citrp
qol: better door sounds
sound: changed sound back to better original airlock sound
code: re enables skyrat's door sound override
/:cl: