Fixes fake lattice linters when they spawn on lava (#1148)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Fixes fake lattice linters when they spawn on lava.

## Why It's Good For The Game

https://www.youtube.com/watch?v=qNFrv0X1W2g

## Proof Of Testing

If it passes linters, it werks.

## Changelog

:cl: BurgerBB
fix: Fixes fake lattice linters when they spawn on lava
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: nevimer <77420409+nevimer@users.noreply.github.com>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>