Replaces the Wizard's High Frequency Blade because I saw Merek get gibbed by it once. (#1081)
## About The Pull Request

Replaces the High Frequency Blade the wizard gets with a REAL banhammer,
and reduces the spell cost from 3 to 2.
Adds REAL banhammers send people to the void temporarily (similar to
immortality talisman or void genetic power) on hit. Hitting someone on
combat mode with it sends them to permabrig instead, if they're not
already in the permabrig.

## Why It's Good For The Game



Lets look at what the High Frequency blade does:
- 100% block chance against projectiles if wielded.
- 50% block chance against melee attacks if wielded.
- 25% block chance otherwise.
- Average case 17 damage per hit, best case 36 damage per hit.
- +20 wound bonus.
- +25 bare wound bonus.
- Attack cooldown of 0.1 seconds (!!!)
- Gibs dead people on hit (very high chance to).
- Deconstructs walls on hit.
- https://www.youtube.com/watch?v=SYUHaf0BxFo

Very insane funny weapon that is way too powerful. This PR replaces that
weapon with a funny banhammer that sends people to perma brig or the
void instead.


## Proof Of Testing


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/25fe06ff-eca3-475c-86d1-258ff00ac8d2)


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/6c152bb1-8684-422b-8c90-428782ca8ca2)


## Changelog

:cl: BurgerBB
add: Adds REAL banhammers send people to the void temporarily (similar
to immortality talisman or void genetic power) on hit. Hitting someone
on combat mode with it sends them to permabrig instead, if they're not
already in the permabrig.
del: Replaces the High Frequency Blade the wizard gets with a REAL
banhammer, and reduces the spell cost from 3 to 2.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>