Dismemberment Alternative (#978)
Alternative to #976

## About The Pull Request

Heads can only be dismembered if they have a critical-tier (Compound)
fracture.
If a head was supposed to be dismembered from a blow, a compound
fracture will be applied.

## Why It's Good For The Game

This is a replacement to the up and coming dismemberment rework that
makes dismemberments harder to obtain through accidents. It is
functionally the same, however with the reduced grossness for it.

The reason why I am so against that PR is that I think it's too gruesome
for SS13. Yes, we have disemboweling, delimbing, gibing, etc etc but
personally I draw the line at still being alive and aware with your
brain exposed. It legitimately gives me nausea and headaches thinking
about it because of a condition I have.

## Proof Of Testing

Untested. Uncompiled. Unhinged.

## Changelog

:cl: BurgerBB
balance: Heads can only be dismembered if they have a critical-tier
(Compound) fracture. If a head was supposed to be dismembered from a
blow, a compound fracture will be applied.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: nikothedude <59709059+nikothedude@users.noreply.github.com>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>