Disables Lance Shuttle on Moon Station because it caused the crew to get stuck on station for 5 hours once. (#1170)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Lance station doesn't work on Moon Station because it's too big. This PR
fixes that by disabling the Lance on Moon Station.

## Why It's Good For The Game

Remove, don't improve, because I do not have the time or patience to
code a system where shuttles are disabled based on the size of the
dockport because that involves a lot of non-modular fuckery. Maybe
someone else can do it.

## Proof Of Testing


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/a6a2f7c0-0e5e-427e-88f2-4a70b4656080)

## Changelog

:cl: BurgerBB
del: Disables Lance Shuttle on Moon Station because it caused the crew
to get stuck on station for 5 hours once.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>