Reverts a few of the security alert sounds back to SR's (#840)
## About The Pull Request

A few of Bubberstation's current security alerts have been reverted back
to Skyrat's (blue, orange, amber, red, gamma).

With the previous (BurgerLUA's) PR (#580)'s main point in mind (this
being, the unnecessary loudness present in most of SR's alert sounds),
each and every .ogg added was refined in Audacity to decrease the peak
without necessarily altering the volume.

## Why It's Good For The Game

Although "quieter" alerts were indeed necessary, some of the media added
didn't match SS13's environment, and most of the time, didn't exactly
make up for the threat which led to the change of alert in the first
place and wasn't capable of maintaining immersion (for example, red
alert being Among Us' imposter sabotage alarm sound while there's a
maniac wielding a double energy sword killing security). Skyrat's are
loyal to the dynamic and authenticity of the game, and certainly live up
to Bubberstation's mildly chaotic rounds.

Attention! For the sake of your ears, the comparison video, because it's
including /both/ loud and quieter audios, it can, obviously, be
uncomfortable to your ears.


https://github.com/Bubberstation/Bubberstation/assets/64568243/f5147211-3c94-457b-89e1-cc95dfe4876b

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

🆑 Chelxox
sound: A few of the alert sounds were reverted back to Skyrat's and
modified to reduce peaks and make them more comfortable to the ear.
/🆑

<!-- Both 🆑's are required for the changelog to work! You can put
your name to the right of the first 🆑 if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>